fix(c++): handle cases when cannot open file in detect
Unfortunately libmagic does not return an error for that case but a
a string starting with 'cannot open ...'. We now tranform that into a
proper error result starting with 'ERROR:'.

Additionally, there was a bug in the detect method implementation
returning always the result from the call the libmagic instead of
returning the transformed result.